+++ contracts/nois-icecube/src/error.rs
@@ -3,6 +3,7 @@ use thiserror::Error;
 use cosmwasm_std::StdError;
 
 #[derive(Error, Debug, PartialEq)]
+#[non_exhaustive]
 pub enum ContractError {
     #[error("{0}")]
     Std(#[from] StdError),